added two limitations on the code snippet's visibility:
A limit on the number of views will allow viewing the snippet only a certain number of times, after which the snippet is deleted from the database.
A limit on the viewing time will allow viewing a code snippet for a certain period of time, and after its expiration, the code snippet is deleted from the database.